Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/xhr/access-control-and-redirects-async-expected.txt


PASS Request is redirected without CORS headers to a response with Access-Control-Allow-Origin=*
PASS Request is redirected to a response with Access-Control-Allow-Origin=*
PASS Request with user info is redirected to a response with Access-Control-Allow-Origin=*
PASS Request is redirect to a bad URL
PASS Preflighted request is redirected to a response with Access-Control-Allow-Origin=*
PASS Preflighted request is redirected to a response with Access-Control-Allow-Origin=* and header allowed
PASS Request is redirected to a same-origin resource file
